requirements.txt
spotipy==2.25.2
streamlit==1.52.2
pandas==2.3.3
matplotlib==3.10.8
python-dotenv>=1.0.0